api/requirements.txt
amqp==2.2.2
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
billiard==3.5.0.3
celery==4.0.2
certifi==2018.1.18
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.4.2
dj-database-url==0.4.1
Django==2.0.1
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.7.7
gunicorn==19.6.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
kombu==4.1.0
msgpack-python==0.5.1
oauthlib==2.0.6
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
requests==2.18.4
shortuuid==0.4.3
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
zope.interface==4.4.3
